chess_analyzer/requirements.txt
python-chess>=1.999
requests>=2.31.0
pandas>=2.0.0
stockfish>=3.28.0
